Move resolveAuth to Service layer.
