+++ src/main/service/index.ts
@@ -6,5 +6,6 @@
  *
  */
 
+export { resolveAuth } from './Auth';
 export { resolveAuthPin } from './AuthPin';
 export { resolveCredential } from './Credential';
